Volume-of-Fluid: fix incorrect sign and add geometric test
The scheme corresponds to calculating the overlap volume of a cell shifted by the flow velocity with its neighbor cell. The new test does that explicitly, but generates convoluted expressions which should not be used directly because they are so slow to evaluate.

Thanks to this test, an incorrect sign in the implementation was found and fixed.